Restructure packages
Move test files to where source files located
Move portgroup_helper.go to helper package
Delete test files in helper and client package for coverage improve
