+++ powermax/provider/host_datasource_test.go
@@ -1,4 +1,4 @@
-package test
+package provider
 
 import (
 	"regexp"
