+++ powermax/provider/masking_view_data_source_test.go
@@ -1,6 +1,6 @@
 // Copyright ©2023 Dell Inc. or its subsidiaries. All Rights Reserved.
 
-package test
+package provider
 
 import (
 	"regexp"
