+++ powermax/provider/storagegroup_resource_test.go
@@ -1,5 +1,5 @@
 // Copyright ©2023 Dell Inc. or its subsidiaries. All Rights Reserved.
-package test
+package provider
 
 import (
 	"testing"
